replace conversations store with context
