+++ ui/package.json
@@ -9,6 +9,7 @@
     "babel-plugin-transform-define": "^1.3.1",
     "dotenv": "^6.1.0",
     "event-source-polyfill": "^1.0.5",
+    "eventsource": "^1.0.7",
     "isomorphic-unfetch": "^3.0.0",
     "js-cookie": "^2.2.0",
     "jss": "latest",
